Adding Sefton Council
First bit of Python (and I'm sure it shows!) so feel free to provide any critiques and if anybody wants to suggest how to make the GUI prompts work better I'm all ears. That said, it does manage to retrieve information from Sefton Council's webpages for bin collections, so it's a (functioning) start!